Move configs to a separate folder
