+++ scripts/build.js
@@ -12,7 +12,7 @@ process.env.NODE_ENV = 'production';
 var path = require('path');
 var rimrafSync = require('rimraf').sync;
 var webpack = require('webpack');
-var config = require('../webpack.config.prod');
+var config = require('../config/webpack.config.prod');
 
 var isInNodeModules = 'node_modules' ===
   path.basename(path.resolve(path.join(__dirname, '..', '..')));
