+++ scripts/start.js
@@ -11,7 +11,7 @@ process.env.NODE_ENV = 'development';
 
 var webpack = require('webpack');
 var WebpackDevServer = require('webpack-dev-server');
-var config = require('../webpack.config.dev');
+var config = require('../config/webpack.config.dev');
 var execSync = require('child_process').execSync;
 var opn = require('opn');
